requirements.txt
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
anthropic>=0.25.0
oandapyV20>=0.6.3
python-dotenv>=1.0.0
